rename table-columns to tableColumns
