+++ ui/lib/apps/ClusterInfo/components/HostTable.js
@@ -8,7 +8,7 @@ import { Tooltip, Typography } from 'antd'
 import React from 'react'
 import { useTranslation } from 'react-i18next'
 import { red } from '@ant-design/colors'
-import { dummyColumn } from '@lib/utils/table-columns'
+import { dummyColumn } from '@lib/utils/tableColumns'
 
 const { Text } = Typography
