+++ ui/lib/apps/SlowQuery/utils/tableColumns.tsx
@@ -17,7 +17,7 @@ import {
   TextWithInfo,
   TextWrap,
 } from '@lib/components'
-import { dummyColumn } from '@lib/utils/table-columns'
+import { dummyColumn } from '@lib/utils/tableColumns'
 
 function ResultStatusBadge({ status }: { status: 'success' | 'error' }) {
   const { t } = useTranslation()
