+++ ui/lib/apps/Statement/components/StatementsTable.tsx
@@ -2,7 +2,7 @@ import React, { useMemo } from 'react'
 import { useNavigate } from 'react-router-dom'
 import { CardTableV2, ICardTableV2Props } from '@lib/components'
 import { StatementTimeRange, StatementModel } from '@lib/client'
-import { statementColumns } from '../utils/table-columns'
+import { statementColumns } from '../utils/tableColumns'
 import DetailPage from '../pages/Detail'
 
 interface Props extends Partial<ICardTableV2Props> {
