+++ ui/lib/apps/Statement/pages/Detail/PlanDetailTabTxn.tsx
@@ -2,7 +2,7 @@ import React from 'react'
 import { StatementModel } from '@lib/client'
 import { CardTableV2 } from '@lib/components'
 import { getValueFormat } from '@baurine/grafana-value-formats'
-import { valueColumns } from '@lib/utils/table-columns'
+import { valueColumns } from '@lib/utils/tableColumns'
 
 export interface ITabTxnProps {
   data: StatementModel
