+++ ui/lib/apps/Statement/pages/Detail/index.tsx
@@ -17,7 +17,7 @@ import { ArrowLeftOutlined } from '@ant-design/icons'
 import { Space, Skeleton, Alert } from 'antd'
 import { SelectionMode } from 'office-ui-fabric-react/lib/DetailsList'
 import { Selection } from 'office-ui-fabric-react/lib/Selection'
-import { planColumns as genPlanColumns } from '../../utils/table-columns'
+import { planColumns as genPlanColumns } from '../../utils/tableColumns'
 import PlanDetail from './PlanDetail'
 import CopyLink from '@lib/components/CopyLink'
 import formatSql from '@lib/utils/formatSql'
